프론트 수정하면서 수정쪽이랑 읽기쪽 수정

+++ src/main/java/com/t1membership/item/dto/ItemDto.java
@@ -7,6 +7,8 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @ToString
@@ -17,7 +19,7 @@ public class ItemDto {
 
     private String itemName;
 
-    private int itemPrice;
+    private BigDecimal itemPrice;
 
     private int itemStock;
